requirements.txt
requests==2.28.2
Django==4.2.5
djangorestframework==3.14.0
pytest==8.0.0
pytest-django==4.8.0